backend/requirements.txt
fastapi
uvicorn
duckdb
pandas
pyarrow
httpx
beautifulsoup4
python-dotenv
loguru
scikit-learn
numpy
tqdm
groq
pyalex
boto3
botocore
